+++ packages/stream_core/lib/stream_core.dart
@@ -1,7 +1,7 @@
 export 'package:dio/dio.dart';
 
 export 'src/api.dart';
-export 'src/attachment.dart';
+export 'src/attachment.dart' hide AttachmentUploadBatchImpl, AttachmentUploadTaskImpl;
 export 'src/errors.dart';
 export 'src/logger.dart';
 export 'src/platform.dart';