+++ packages/stream_core/test/api/interceptors/auth_interceptor_test.dart
@@ -1,6 +1,5 @@
 import 'dart:async';
 import 'dart:convert';
-import 'dart:typed_data';
 
 import 'package:stream_core/stream_core.dart';
 import 'package:test/test.dart';